chore: Remove deprecated API handler and moved that to an app router … (#444)
