+++ apps/earn-protocol/constants/sdk.ts
@@ -1,0 +1,1 @@
+export const sdkApiUrl = '/api/sdk'
