--- apps/earn-protocol/providers/SDK/index.ts
@@ -1,4 +1,0 @@
-export * from './context'
-export * from './deposit'
-export * from './withdraw'
-export * from './prepareTransaction'
